coming, here, moving done

+++ demo/src/main/kotlin/com/example/demo/repo/OrderStatusRepoImpl.kt
@@ -1,7 +1,11 @@
 package com.example.demo.repo
 
 import com.example.demo.model.*
+import jakarta.servlet.ServletException
 import org.springframework.data.mongodb.core.MongoTemplate
+import org.springframework.data.mongodb.core.query.Criteria
+import org.springframework.data.mongodb.core.query.Query
+import org.springframework.data.mongodb.core.query.isEqualTo
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Component
 import org.springframework.transaction.support.TransactionTemplate
@@ -17,18 +21,23 @@ class OrderStatusRepoImpl(
         val upcomingRepoImpl: UpcomingRepoImpl,
         val orderHereRepoImpl: OrderHereRepoImpl,
         val movingRepoImpl: MovingRepoImpl,
-        val serviceAddressRepo : ServiceAddressRepo
+        val serviceAddressRepoImpl : ServiceAddressRepoImpl
 ) : OrderStatusRepo {
     override fun addOrderStatus(orderStatus: OrderStatus): OrderStatus {
         return mongoTemplate.save(orderStatus)
     }
 
     override fun updateStatus(orderId:String, serviceAddressId: String): OrderStatus {
+        val listStatus = getListStatusById(orderId)
+        val orderStatus : OrderStatus = listStatus[listStatus.size-1]
         return transactionTemplate.execute{_ ->
 
-            val orderStatus = getStatusById(orderId)
-            val serviceAddress = serviceAddressRepo.getServiceAddressById(serviceAddressId)
+            println("orderrrr iddd " + orderId)
+            println(orderStatus)
+//            val orderStatus = getStatusById(orderId)
+            val serviceAddress = serviceAddressRepoImpl.getServiceAddressById(serviceAddressId)
             val updatedStatus = orderStatus.update(serviceAddress)
+            println(updatedStatus)
             val newOrder = orderRepo.orderById(orderId)
 
             if(updatedStatus.statusNum == 2) {
@@ -36,8 +45,9 @@ class OrderStatusRepoImpl(
                 movingRepoImpl.deleteMove(orderId)
                 orderHereRepoImpl.createOrderHere(OrderHere(serviceAddressId, newOrder))
             }
-            else {
+            if(updatedStatus.statusNum == 1) {
                 val serviceAddressIdAfter = orderHereRepoImpl.findIdHere(orderId)
+                println("ok " + serviceAddressIdAfter)
                 upcomingRepoImpl.createUpcoming(Upcoming(serviceAddressId, newOrder))
                 movingRepoImpl.createMoving(Moving(serviceAddressIdAfter, newOrder))
                 orderHereRepoImpl.deleteHere(orderId)
@@ -51,14 +61,32 @@ class OrderStatusRepoImpl(
         }
 
     override fun getStatusById(orderId: String): OrderStatus {
-        val listStatus : List<OrderStatus> = mongoTemplate.findAll(OrderStatus::class.java)
-        for(orderStatus : OrderStatus in listStatus)
-            if(orderStatus.orderId == orderId)
+//        val query = Query()
+//        query.addCriteria(Criteria.where("orderId").isEqualTo(orderId))
+//        val listStatus = mongoTemplate.find(query, OrderStatus::class.java)
+        val listStatus = mongoTemplate.findAll(OrderStatus::class.java)
+//        var index = listStatus.size - 1
+//        while (index >= 0){
+//            val orderStatus = listStatus[index]
+//            if(orderStatus.orderId == orderId){
+//                println("id " + orderId + " true")
+//                return orderStatus
+//            }
+//            index = index - 1
+//        }
+        for(orderStatus in listStatus){
+            println("order id " + orderStatus.orderId)
+            println(orderId)
+            if(orderStatus.orderId == orderId){
+                println("id " + orderId + " true")
                 return orderStatus
-        throw ResponseStatusException(
+            }
+        }
+                throw ResponseStatusException(
                 HttpStatus.NOT_FOUND,
-                "Cannot find any event with id $orderId"
+                "Cannot find any status with order id"
         )
+
 
 //        return mongoTemplate.findById(orderId, OrderStatus::class.java)?:throw ResponseStatusException(
 //                HttpStatus.NOT_FOUND,
